More migration to ARC.

+++ PreferencesController.h
@@ -244,16 +244,16 @@ extern NSString* SynchronizationAdditionalScriptsKey;
 -(NSArrayController*) editionTextShortcutsController;
 
 -(NSArray*)             preambles;
--(int)                  preambleDocumentIndex;
--(int)                  preambleServiceIndex;
+-(NSInteger)            preambleDocumentIndex;
+-(NSInteger)            preambleServiceIndex;
 -(NSAttributedString*)  preambleDocumentAttributedString;
 -(NSAttributedString*)  preambleServiceAttributedString;
 -(PreamblesController*) preamblesController;
 
 -(NSArray*)                 bodyTemplates;
 -(NSArray*)                 bodyTemplatesWithNone;
--(int)                      bodyTemplateDocumentIndex;
--(int)                      bodyTemplateServiceIndex;
+-(NSInteger)                bodyTemplateDocumentIndex;
+-(NSInteger)                bodyTemplateServiceIndex;
 -(NSDictionary*)            bodyTemplateDocumentDictionary;
 -(NSDictionary*)            bodyTemplateServiceDictionary;
 -(BodyTemplatesController*) bodyTemplatesController;
@@ -262,8 +262,8 @@ extern NSString* SynchronizationAdditionalScriptsKey;
 -(NSArray*)           compositionConfigurations;
 -(void)               setCompositionConfigurations:(NSArray*)value;
 
--(int)                compositionConfigurationsDocumentIndex;
--(void)               setCompositionConfigurationsDocumentIndex:(int)value;
+-(NSInteger)          compositionConfigurationsDocumentIndex;
+-(void)               setCompositionConfigurationsDocumentIndex:(NSInteger)value;
 -(NSDictionary*)      compositionConfigurationDocument;
 -(void)               setCompositionConfigurationDocument:(NSDictionary*)value;
 
@@ -286,7 +286,7 @@ extern NSString* SynchronizationAdditionalScriptsKey;
 
 -(BOOL)                      encapsulationsEnabled;
 -(NSArray*)                  encapsulations;
--(int)                       encapsulationsSelectedIndex;
+-(NSInteger)                 encapsulationsSelectedIndex;
 -(NSString*)                 encapsulationSelected;
 -(EncapsulationsController*) encapsulationsController;
 
@@ -307,8 +307,8 @@ extern NSString* SynchronizationAdditionalScriptsKey;
 -(NSDictionary*)                               synchronizationAdditionalScripts;
 -(SynchronizationAdditionalScriptsController*) synchronizationAdditionalScriptsController;
 
--(int)    paletteLaTeXGroupSelectedTag;
--(void)   setPaletteLaTeXGroupSelectedTag:(int)value;
+-(NSInteger)    paletteLaTeXGroupSelectedTag;
+-(void)   setPaletteLaTeXGroupSelectedTag:(NSInteger)value;
 -(NSRect) paletteLaTeXWindowFrame;
 -(void)   setPaletteLaTeXWindowFrame:(NSRect)value;
 -(BOOL)   paletteLaTeXDetailsOpened;
